% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.CER.lrm.R
\name{calc.CER.lrm}
\alias{calc.CER.lrm}
\title{Calculate effect sizes for CER analyses}
\usage{
calc.CER.lrm(mids.data = NULL, orig.data = NULL, instrument = NULL,
  therapy = NULL, outcome = NULL, subset.cer = NULL)
}
\arguments{
\item{mids.data}{A MIDS: a MICE imputed dataset}

\item{orig.data}{The original data.frame/data.table which was imputed}

\item{instrument}{A character of the name of the instrumental variable}

\item{therapy}{A character of the name of the therapy}

\item{outcome}{A character of the outcome parameter}

\item{subset.cer}{A logical vector (with length equal to the number of cases), which indicates included and excluded cases}
}
\value{
Returns a vector of the estimate with 95% CI of the treatment effect, including the R2 of the stage 1 regression.
}
\description{
This function uses two-stage least squares regression to estimate a causal treatment effect with an instrumental variable. For info or explanation, please contact: b.gravesteijn@erasmusmc.nl
}
