% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.boot.R
\name{cluster.boot}
\alias{cluster.boot}
\title{Bootstrap with clustered data}
\usage{
cluster.boot(data = NULL, cluster.v = NULL, statistic = NULL, R = NULL)
}
\arguments{
\item{data}{data.frame or data.table}

\item{cluster.v}{the grouping variable of interest}

\item{statistic}{the function to use on the bootstrap sample}

\item{R}{the number of bootstrap replicates}
}
\description{
Bootstrap with clustered data
}
