% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV.calc.R
\name{IV.calc}
\alias{IV.calc}
\title{Instrumental variable analysis for multiple imputed datasets.}
\usage{
IV.calc(instrument = NULL, therapy = NULL, outcome = NULL,
  mids = NULL)
}
\arguments{
\item{instrument}{A character value of the name of the instrumental variable (any type)}

\item{therapy}{A character value of the name of the binary therapy}

\item{outcome}{A character value of the name of the outcome variable, both binary as ordinal.}

\item{mids}{A mids object, of the multiple imputed dataset which we will use for the IV analysis.}
}
\value{
Returns a vector with the effect estimate with 95% CI.
}
\description{
Using an instrumental variable approach, calculate the effect estimate for a therapy. As instrument, any type of instrument may be used (continuous/ordinal/binary/categorical). The therapy must be a binary variable, and the outcome may be both binary (logistic regression) as ordinal (proportional odds regression). Make sure the assumptions are correct, before using this funciton!
}
