% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOR.ci.calc.R
\name{MOR.ci.calc}
\alias{MOR.ci.calc}
\title{Calculate the median odds ratio and CI using bootstrap}
\usage{
MOR.ci.calc(formula = NULL, data = NULL, grp.var.t = NULL)
}
\arguments{
\item{formula}{a character containing the formula for the glmer function}

\item{data}{a (list of) data.frame(s) or data.table(s), or a MIDS object
containing the data on which the model was fitted}

\item{grp.var.t}{the grouping variable, e.g.: "country"}
}
\description{
Calculate the median odds ratio and CI using bootstrap
}
