% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORcalc.R
\name{MORcalc}
\alias{MORcalc}
\title{Calculate the median odds ratio}
\usage{
MORcalc(my.var, digits = 2)
}
\arguments{
\item{my.var}{= variance associated with level 2 clustering variable}

\item{digits}{= number of decimal places to which MOR value will be rounded.}
}
\description{
Calculate the median odds ratio
}
