% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Make a caterpillar plot for random effects}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot object}

\item{cols}{number of columns}

\item{layout}{matrix describing how the plots should be ordered}
}
\value{
Returns a plot containing all ggplot objects
}
\description{
Make a caterpillar plot for random effects
}
