% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nagelkerke.R
\name{Nagelkerke}
\alias{Nagelkerke}
\title{Calculate Nagerkere R2 for fitted models}
\usage{
Nagelkerke(fit = NULL, nullmod = NULL, data = NULL)
}
\arguments{
\item{fit}{a (list of) fitted model(s), which need to be compared to the null model}

\item{nullmod}{a null model, to which the fitted model(s) need to be compared}

\item{data}{the dataset on which the model is fitted}
}
\value{
the Nagelkerke R2, as a numer between 0 and 1
}
\description{
Calculate Nagerkere R2 for fitted models
}
