% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probplot.R
\name{probplot}
\alias{probplot}
\title{Compare proportions in categorical variables between observed and imputed data
visually (using ggplot)}
\usage{
probplot(x, formula, facet = "wrap", ...)
}
\arguments{
\item{x}{x: mids object (from mice)}

\item{x}{formula: formula describing which variables to plot}

\item{x}{facet: either "wrap" for facet_wrap or "grid" for facet_grid}

\item{x}{...: additional parameters passed to theme()

Note: if the formula is not specified, all imputed categorical variables are
plottet. A formula has the structure
categorical variables ~ faceting variables | color variable
By default, .imp (imputation set identifier) will be used as color variable}
}
\description{
Parameters:
}
