% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC plot.R
\name{roc.model}
\alias{roc.model}
\title{Make a ROC plot, with AUC}
\usage{
roc.model(x, obs.v)
}
\arguments{
\item{x}{Alist of fitted models (for multiple imputed dataset fitted models)}

\item{obs.v}{A vector of the observed outcome variable: (0,1)}
}
\value{
Returns a ROC plot containing AUC
}
\description{
Make a ROC plot, with AUC
}
