% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catterpillar.R
\name{catterpillar}
\alias{catterpillar}
\title{Make a caterpillar plot for random effects}
\usage{
catterpillar(x = NULL, fitter = NULL, grp.var.t = NULL, plotMOR = TRUE,
  xMOR = 3, yMOR = 2)
}
\arguments{
\item{x}{A (list of) fitted model(s) (for multiple imputed dataset fitted models)}

\item{fitter}{A character indicating random effects fitting formula, e.g: "glmer"}

\item{grp.var.t}{character indicating type of grouping variable, e.g.: "country"}

\item{xMOR}{x coordinate of the text of MOR (default is 3)}

\item{yMOR}{y coordinate of the text of MOR (default is 2)}

\item{printMOR}{logical indicating whether the median odds ratio should be printed, default is TRUE}
}
\value{
Returns a plot containing mean and 95% CI, with pooled conditional variances using Rubin's rules.
}
\description{
Make a caterpillar plot for random effects. Please use carefully, it bites.
}
