% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.na.plot.R
\name{distr.na}
\alias{distr.na}
\title{Distribution and NA plot}
\usage{
distr.na(x = NULL)
}
\arguments{
\item{x}{An object with data}
}
\value{
Returns a plot showing distribution and missingness of variables
}
\description{
Distribution and NA plot
}
